+++ CHANGELOG.md
@@ -60,3 +60,4 @@ Initial version
   * added support for dataTraceConf
   * added support for cpuSpeed element in serial config (required for swo mode in serial tracing)
   * automatically ceil duration of generalConf since only integers are allowed
+  * fixed error message for gdbPort